Make sure matching on '/' will work
It's possible that someone would want to add a path prefix of just
'/'. With the current layout there isn't a good way to have a leaf node
for that path, as leaves all have some meaningful path component.

Fix that by adding a dummy /root to all paths when
adding/searching. For a path of '/', you can hold a value below that
dummy leaf node. For all other paths, the real paths start below that
dummy node.