Prefix search path components (#1014)
* Change prefix_search_... to use list of path parts

Add the ability to use prefix_search_.. with a list of already-split
path components instead of a path as string/filter_value_t. That allows
custom use where the string has different components (like http url
proto/domain/port/path, etc), but you still want to allow parallel
searching on the components.

This passes the unit tests so far although I'm going to expand them.

* Make sure matching on '/' will work

It's possible that someone would want to add a path prefix of just
'/'. With the current layout there isn't a good way to have a leaf node
for that path, as leaves all have some meaningful path component.

Fix that by adding a dummy /root to all paths when
adding/searching. For a path of '/', you can hold a value below that
dummy leaf node. For all other paths, the real paths start below that
dummy node.