Fix AT_FDCWD 32-bit syscall decoding
AT_FDCWD is equal to -100. In the 64-bit version of the syscall, its
value as unsigned long is 18446744073709551516 (1<<64 - 100). However,
when the 32-bit syscall version is used, the value is truncated to 32
bits and AT_FDCWD is passed as 4294967196 (1<<32 - 100).

Cast the value to int to drop the high 32 bits (the dirfd argument is
defined to be an int anyway) and then compare the signed 32-bit values.