Add support for bpf() syscall
The return value res in the bpf exit event is dynamic. On failure or
when the bpf command is something other than a program/map load, it's
considered an errno value.

On success and when the command is a program/map load, it's the fd from
which the data from the program/map can be read.