Allow "in" operator to work with non-string values (#1049)
When used by falco, an expression like "xxx in (a, b, c)" is stored with
an operator CO_IN and the values a, b, c, etc as values added via calls
to add_filter_value. (In sysdig, it's a bit more nuanced, with the
treatment dependent on what the type of xxx is).

When checking, a the values a, b, c are stored in a set and a set
membership test tells you if a given test value matches one of a, b, c.

For non-string values, The values a, b, c, etc are parsed into raw
values, converting things like numeric strings to 16 bit values, booleans
to 0/1, etc.

This had a bug, though, as the value being stored in the set has its
length set to the length of the original value (e.g. the string
representing the port), and not the parsed raw value (2 bytes).

This fixes that bug by storing the parsed raw length instead of the unparsed
length. In turn, that required some other changes to return the parsed
length from functions like sinsp_filter_check::parse_filter_value and
sinsp_filter_value_parser::string_to_rawval.