Add a filtercheck for when a fd name changes
Add a filtercheck fd.name_changed which is true when parsing an event
changes the "name" of a fd. This mostly occurs when a
bind/connect/sendto/recvfrom/etc changes the ip/port information
associated with a socket fd, and is useful if you want to track when a
stream of sendto/recvfroms on a single socket changes addresses.

This is done by adding a m_oldname to fdinfo, which is set when
threadinfo returns a fd and associates it with an event. If after
parsing, the name changes, the filtercheck will return true.

Whether the name changed or not is an event property and set at the end
of sinsp_parser::process_event(). It does this by checking the original
fd name and comparing it to the fd name that exists after parsing the
event.