Make block type 7 the current version for captures
https://github.com/draios/sysdig/pull/1044 created a new block type
PL_BLOCK_TYPE_V8 that stored the pgid in capture files.

We want that block type to be supported in open source sysdig, and to be
known to the agent, but not generated by the agent yet. So we'll change
libscap to keep the current block number at V7 for now and to write
blocks with the format used by V7 instead of V8.

A later PR will undo this change, causing the driver to generate
captures with block type V8. At that time, open source sysdig will be
widely available and users can read the captures.