Downgrade execve version to 18 in driver
https://github.com/draios/sysdig/pull/1044 changed the event number for
execve events from 18 to 19, in order to add the pgid to the exit event.

We want that event to be supported in open source sysdig, and to be
known to the agent, but not generated by the agent yet. So we'll change
the driver event table to use version 18 for execve events instead of
19, and to return only those fields present in execve version 18.

A later PR will undo this change, causing the driver to generate execve
events with version 19. At that time, open source sysdig will be widely
available and users can read the events in captures.