Change setpgid parsing to work w/ namespaced args
Setpgid()s arguments are from the namespace of the caller. So in some
cases, the pid used as an argument to setpgid() doesn't directly map to
any threadinfo.

To fix this, change event parsing to determine whether the caller is
namespaced or not. In a namespace, first look up the corresponding pid
in the global namespace before looking up the threadinfo.