Allow capture files with versions < software
The test between capture file version and software version was doing a
strict comparison--the capture would only be read if the two versions
were equal. This prevents newer versions of sysdig from reading older
capture files.

To fix this, change the test to only refuse to read a capture if its
major number is greater or major numbers are the same but minor number
is greater.