Fix pgid ns (#1080)
* Rename pgid to vpgid to reflect it's in-namespace

https://github.com/draios/sysdig/pull/1044 added tracking of a process's
process group id (pgid). However, this change didn't compeltely handle
process group ids in a namespace. When reading process information from
/proc, it was using the process group id from the global namespace. When
tracking execve()s and setpgid()s it would take the pgid from the
namespace of the calling process.

To fix this, define the pgid to be the pgid from its current pid
namespace. And to make this clear, rename it to vpgid everywhere.

This commit handles renaming the variable at the scap/sinsp
levels. Other changes will fix setpgid() handling to handle cases when
it's called in a pid namespace.

* Rename the proc.pgid filtercheck to proc.vpgid

This reflects its real meaning (from the namespace)

* Removing event parsing for setpgid

This means that if a setpgid occurs, the pgid of the process won't be
in-sync again until the next execve(), where the process's pgid is
returned by the driver in the exit event. However, we expect that this
will cover most cases, and is dramatically simpler.

* Fix logic for setting vpgid

Comment was right, but test was backwards.