Add prctl, update threadinfo on prctl(PR_SET_NAME)
Add support for prctl events. prctl is a fairly generic syscall with
general-purpose arguments which change interpretation depending on the
first option argument.

This does minimal filling, generally passing the arguments as-is. One
exception is in the specific case of PR_SET_NAME, where the arg2
argument is a char * with the new program name. For that option, the
type is PT_DYN and changes from an unsigned long to PR_CHARBUF depending
on the value of option.

In userspace, the parser will update a threadinfo's m_comm upon prctl
exit, assuming the result is successful and if the option is PR_SET_NAME.