Allow fd.port to be used with in.
When the filtercheck is fd.port, it's handled by
sinsp_filter_check_fd::compare_port, which checks both the source and
destination port. This didn't accept an operator of CO_IN.

Change this to do flt_compare() using the source and destination port,
whichh in turn uses set membership tests.