Allow evttype filters to work with syscalls (#1100)
* Allow evttype filters to work with syscalls

Syscalls have their own numbers but they weren't really handled by
sinsp_evttype_filter. This meant that there wasn't a way to handle
filters with evt.type=xxx clauses where xxx was a value that didn't have
a corresponding event entry (like "madvise", for examples), or where a
syscall like open could also be done indirectly via syscall(__NR_open,
...).

This fixes that by also tracking the syscall numbers for an event like
"open" or "madvise". They are passed to add() as a separate set and
there are corresponding arrays/lists that map from a syscall number to a
list of filters to run related to that syscall.

in ::run(), if the event number if PPME_GENERIC_E/X, look at the
argument to find the syscall number and index into the syscall-specific
vectors to find the right filters.

* Move filters assignment to own branch.

Makes it more clear about when it's set based on evttype vs syscall.