edit Sysdig license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the COPYING file so that it contains only the full
text of the GPL 2.0 license. The info that pertains to OpenSSL has now
been moved to the "License Terms" section in the README.

Collectively, these changes allow Licensee to successfully identify the
license type of Sysdig as GPL 2.0.

sysdig-CLA-1.0-contributing-entity: Cloud Native Computing Foundation
sysdig-CLA-1.0-signed-off-by: Andrea Kao <eirinikos@gmail.com>