Flag filters using args with EPF_REQUIRES_ARGUMENT
Many filters required an argument in the form .arg or [arg], but they
didn't have the flag EPF_REQUIRES_ARGUMENT. This flag wasn't used
anywhere, but we plan on using it in falco soon, so it needs to be
accurate.

Do a pass over all filterchecks that call extract_arg and add
EPF_REQUIRES_ARGUMENT as needed.