Change sinsp_container_manager::get_container to return a ptr
Change sinsp_container_manager::get_container() to return a pointer
instead of filling in an object. There was a private version already
that did this, so modify it slightly to return a const ptr.

We found that this could affect performance when running lots and lots
of events against container-based filters.

This required lots of related changes whenever get_container() was
called, and some follow-on changes to enforce const-ness.