Change thread_to_scap to do a shallow copy
A big cost in creating a new capture file from a live event stream is
dumping the current set of threads to the capture file. This is done
in sinsp_thread_manager::dump_threads_to_file() and involves iterating
over the set of threads and converting each to a scap_threadinfo
struct, multiple times in fact.

To help make that more efficient, we can do a shallow copy of the
strings from the sinsp_threadinfo struct given that the corresponding
scap_threadinfo is temporary.

{args,env,cgroups}_to_scap are replaced by {args,env,cgroups}_to_iovec,
which assembles an iovec containing the strings in the various string
vectors, without actually copying anything in the common case. In the
case of truncation, a "remainder" string is used, which does get a
truncated copy of the last argument/env var/cgroup value that doesn't
fit.

Also, there's now no need to allocate and populate a scap_threadinfo
struct when you're computing the write size for a thread. You can get
all the sizes directly from the sinsp_threadinfo object.