Add way to dump scap_threadinfo w/ alternate bufs
Add a new scap_write_proclist_entry_bufs which takes a scap_threadinfo
and a set of strings representing the various array-backed strings that
can be in a scap_threadinfo struct.

In the case of args/env/cgroups, which are represented within sinsp as
vectors of strings, allow writing as an iovec, so the vector elements
can be written individually without having to be combined first.

This is useful when dumping the set of (sinsp) threads, as you can dump
a thread without copying all of the strings into the scap_threadinfo
struct.