Improve state switches (#1121)
* Add way to dump scap_threadinfo w/ alternate bufs

Add a new scap_write_proclist_entry_bufs which takes a scap_threadinfo
and a set of strings representing the various array-backed strings that
can be in a scap_threadinfo struct.

In the case of args/env/cgroups, which are represented within sinsp as
vectors of strings, allow writing as an iovec, so the vector elements
can be written individually without having to be combined first.

This is useful when dumping the set of (sinsp) threads, as you can dump
a thread without copying all of the strings into the scap_threadinfo
struct.

* Change thread_to_scap to do a shallow copy

A big cost in creating a new capture file from a live event stream is
dumping the current set of threads to the capture file. This is done
in sinsp_thread_manager::dump_threads_to_file() and involves iterating
over the set of threads and converting each to a scap_threadinfo
struct, multiple times in fact.

To help make that more efficient, we can do a shallow copy of the
strings from the sinsp_threadinfo struct given that the corresponding
scap_threadinfo is temporary.

{args,env,cgroups}_to_scap are replaced by {args,env,cgroups}_to_iovec,
which assembles an iovec containing the strings in the various string
vectors, without actually copying anything in the common case. In the
case of truncation, a "remainder" string is used, which does get a
truncated copy of the last argument/env var/cgroup value that doesn't
fit.

Also, there's now no need to allocate and populate a scap_threadinfo
struct when you're computing the write size for a thread. You can get
all the sizes directly from the sinsp_threadinfo object.