Scap process blacklist (#1139)
* libscap: add way to ignore all evts below a comm

Add the ability to ignnore events for a process as well as all of its
children by comm. On fork/clone/exec, the comm return value is checked
and if the comm is found in a list of suppress comms, the tid is added
to a hash table of suppressed tids m_suppressed_tids.

In addition, on clone/fork/exec if the parent tid is in the suppressed
tid set the new tid is always added to the set as well.

On procexit the tid is always removed from the set of suppressed
tids. The procexit is only passed up if the thread was not previously
suppressed.

In scap_stats, include the number of suppresed threads. Will be useful
for identifying runaway groups of suppressed threads in the face of
dropped events and pid recycling, where you might end up with stale pids
in the suppressed tid set.

* libsinsp: add ability to suppress events by comm

Add the ability to suppress events by comm. sinsp::suppress_methods_comm
adds comms to the set. These comm values are passed as arguments to
scap_open(). After scap_open(), additional values are passed to libscap
using scap_suppress_events_comm(). You can not specify more than
SCAP_MAX_SUPPRESSED_COMMS entries.

* Sysdig: add ability to suppress events by comm

New argument -U/--suppress-comm adds comm values to the inspector via
sinsp::suppress_events_comm(). If this fails, return an error and exit.