Update for proto_ops.getname() prototype changes in Linux 4.17 (#1114)
Linux 4.17 removed the sockaddr_len argument and instead returns
the length on success instead of returning 0. To cope, define an
internal wrapper that uses the appropriate prototype based on
kernel version. Since nothing uses the sockaddr make the wrapper
return 0 on success or a negative error code so that surrounding
code does not need to change.