Don't skip name=(...) cgroups
They're useful for custom containers and apparently systemd
only sets the `name=systemd` cgroup right after reboot,
at least on CentOS 7.

Subsequent restarts set all the cgroups to the slice but that
would require restarting all services after boot.

Replace the = character we were using to separate cgroup name
from value with : (it's already used as a separator in
/proc/*/cgroup and we don't really need anything special
from this character, except for not appearing in cgroup names.