Allow k8s filterchecks with analyzer
Allow k8s filterchecks when compiling with the analyzer. This matches
what we do for mesos--the filterchecks may not actually match events, as
we don't directly connect the inspector to external mesos/k8s servers,
but they will at least be defined.

This also ensures that a falco rules file validated with open source
falco will also be validated by the sysdig agent.

As the analyzer does not have the necessary backing information for some
previously ifdef'd out filterchecks, when HAS_ANALYZER is defined
replace these with stubs that just return NULL/empty lists.