Handle ipv6 addrs in comparisons/string conversion
Handle PT_IPV6ADDR as a type in string conversions to/from rawvals, both
directly and into the lua stack. Generally it's just inet_pton/inet_ntop
on the bytes in the ipv6addr struct.

Also handle comparisons on PT_IPV6ADDR/PT_IPV6NET, using the equality
and netmast comparison operators on the struct.

To handle conversion/comparisons for PT_IPADDR/PT_IPNET, use the size of
the rawval to determine if it's an ipv4 or ipv6 address and call
comparison/converson method again with PT_IPV6ADDR/PT_IPV4ADDR as
appropriate. For strings, infer the type based on whether the string has
'.' or ':' characters.

This required a change to the methods to take the param
type/format instead of a filtercheck_field_info, as there are other uses
that don't have a filtercheck_field_info.