Fix fd.net comparisons with in operator
When we added ipv6 support, we created a new param type PT_IPNET which
represents either an ipv4 netmask or ipv6 netmask, with the size being
used to determine the ultimate type.

That means we need to account for that type when using a CO_IN operator
and switch to checking each value in turn instead of doing a set
memberhip test.

This also means we need to call flt_compare with the correct argument
sizes so the size can be used to determine whether the netmask is ipv4
or ipv6.