Use curl_multi_wait()
Don't call curl_multi_perform() in a busy wait loop, since it generates 1M+
system calls on a typical system where we have to wait 300-500ms for Docker
to send the response, leading to high CPU usage and saturation of kernel
buffers. Instead, use curl_multi_wait() with infinite timeout, which
reduces the system calls generated to a dozen.