Add option to conditionally build with -Werror
The -Werror flag is too conservative for general purpose use.  Different
users may have diffent compilers with different warnings.  What compiles
for us with our version of the compiler might not compile for someone
else with a different compiler.

This version makes building conditional, based on a
BUILD_WARNINGS_AS_ERRORS option, which is disabled by default.

This also cleans up the indentation in CMakeLists.txt to consistently
use tabs.