eBPF support for RHEL 7.6
The only unfortunate casualty was switching to populating the cgroup
subsystem name at compile time rather than runtime, which makes the eBPF
state increase significantly. So I removed two more cgroups from the
supported list. It can potentially be reworked to account that, but as long
as we grab one relevant we should be fully able to do the container
detection. It'll probably be worth adding all of them at some point.