Add prototype for async_metric_source
This is a prototype for an abstract asynchronous metric source.  The
idea is that for each metric source (e.g., docker), we'll have a
concrete subclass.  The concrete subclasses will implement the
run_impl() method, which will get invoked in the context of a separate
thread.

Metrics can be returned to the client code in one of three ways:

  (1) If we can fetch the metrics within a specified time window, we'll
      return them synchronously.
  (2) If we don't fetch the metrics fast enough, and if the client code
      provided a callback function, then we'll invoke the callback as
      soon as the metrics are available.
  (3) If we don't fetch the metrics fast enough, and if the client code
      did not provide a callback function, then we'll save the metrics and
      the next time the client code tries to fetch metrics for an entity
      with the same key, we'll return the previously fetched metric.

Option (3) leaves some room for leaking entries -- if we collect metrics
for some entity, store them, but the client code never asks for them
again.  If we can eliminate option (3) and support only options (1) and
(2), then this problem goes away.  Otherwise, I'll need to build in some
way to trim "old" metrics.