Detect container health checks (#1281)
* Track container health check processes

Track and identify processes that are associated with container health
checks:

- When fetching container info, also parse the contents of <root
  container obj>[Config][Healthcheck]. Set a bool m_has_healthcheck to
  true if a healthcheck is configured, and save the exe and args in
  m_healthcheck_exe/m_healthcheck_args. A normalization function
  normalize_healthcheck_arg removes pairs of leading " or ' chars.

- Threadinfo now has a bool m_is_container_healthcheck (default
  false). A method identify_healthcheck(tinfo) sets the flag=true if the
  thread is running in a container that has a healthcheck, the exe/args
  match that healthcheck, it's not the initial process for the
  container (vpid==1), and the thread's ancestors don't include the
  original container init process. It will *not* set it to false by
  default, so a threadinfo that is initially identified as a health
  check will remain one across execs e.g. "sh -c /bin/true" execing
  /bin/true.

* Filterchecks to display/check container healthchks

container.healthcheck shows the configured healthcheck, which is either
"NONE" (configured, but no health check) or the exe + args.

proc.is_container_health_check is true if the threadinfo is associated
with a container health check.

* Save healthcheck info to capture files

Capture files contain a "container event" (json blob) that is used to
hold container information when a container is created. Add a
"Healthcheck" property containing the healthcheck information fetched
while querying the container info.

When reading container events, parse the healthcheck info using the
existing parse_healthcheck().

* Make populate_cmdline part of sinsp_threadinfo

It's useful to use externally in things like unit tests.