Add async Docker metadata collection
This change extends the original framework with the addition of async
Docker metadata collection.  Part of this change is in the same sprit as
a change from Greg -- I've trying to split out the Linux/Windows/Noop
implementations rather than having macro guards throughout the
implementation.  My approach differs a little in the sense that I
introduced a class hierarchy to deal with it.

Note that this is just a prototype; there's a good bit of work that
isn't here.  That includes:
    * Need to implement the Windows version of the collector
    * Need to implement unit tests.  That will include the need to
      implement something that will "act like" Docker.