Introduce url_fetcher
This change pulls the URL fetching logic out of the implementation of
async_linux_docker_metadata_source and into a separate class.  The
rationale for this change is:

    * The details of URL fetching is a separate concern
    * The URL fetching logic might be reusable by other components

The URL fetcher is separated into an interface and a concrete
realization of that interface, with factory methods for creating
concretions.  The rationale for this design is:

    * It avoids class-to-class coupling between
      async_linux_docker_metadata_source and a concrete url fetcher
    * It will enable us to tweak the factory methods in a unit test
      environemnt to return UT-specific realizations of the interface
      that don't actually have to talk to web servers.