Prune metadata cache
The async_docker_metadata_source will cache metadata lookups if the
lookup takes longer than the client is willing to wait.  A subsequent
call to lookup with the same key will returned the cached result.  The
risk is, if nothing ever makes that subsequent call then that cached
result will sit in the cache forever.  Over time, the cache will
continue to leak such metadata.

This change introduces a time to live for cache entries.  When a cache
entry is created, a timestamp is recorded.  The async thread examines
the entries in the cache and prunes ones that are too old, as specified
by the client code.