Reset thread container_id _before_, not _after_ container detection
This allows container engines to return `false` but still set
a container_id, which is useful for Docker and CRI (they may detect
a container but fail to fetch its metadata, at least temporarily).
In that case, we don't want to add a container object (so that a future
lookup may succeed) but still record this process as a part of that
container.