Revert change to make socket NEVER_DROP
After some experiemntation I discovered that making the socket() call
NEVER_DROP didn't help much, because under heavy syscall load we'll can
drop the syscall used to write to the file descriptor.  Rather than put
extra load on the system and force socket() not to be dropped, I'm
reverting to the original approach.