Infer fd info for sendto system call [SMAGENT-1282] (#1304)
* Infer fd info for sendto system call

If we don't have file descriptor information when processing the sendto
system call, then we can't analyze the data associated with the system
call.

This change introduces code to try to infer the file descriptor
information.  We know that the sendto system call will be used with a
network socket, and that it will almost always be used with a UDP
socket.  If the information is missing, we inject that information into
the SENDTO_E event.