Handle mmap failure gracefully
For some reason, scap uses MAP_FAILED as a sentinel to indicate the absence
of the buffer, whereas in the BPF code I was mostly using NULL. So, in the
rare cases where mmap fails, this can cause a segfault while scap_close
cleans up structures.

It would be cleaner to just move all the code to use NULL, but since a
large portion of the kmod-related scap functions rely on the MAP_FAILED,
it's safer to just adapt the BPF part to use MAP_FAILED as well.

As part of the fix, make sure we don't leak any fds when mmap fails.