Misc bug fixes
Ensure that the threadinfo for the container event is a valid running
thread. This means we don't need to pass along a request thread--we'll
keep track of the thread for the container and add it at the time the
container json event is created.

Save the metadata deadline for containers along with other info.