Create a grpc_channel_registry for all channels
Create a new top-level static method
libsinsp::grpc_channel_registry::get_channel(url), backed by a static
map from url to grpc::Channel objects.

This ensures that when creating channels there is only a single channel
for a given url. It appears that if multiple channels are created for a
single url, the underlying grpc implementation may not actually refresh
the connection to each fd (unix socket, connection) if the server goes
away.